requirements/heroku.txt
-r prod.txt
<<<<<<< HEAD
Flask-SSLify==0.1.4
gunicorn==18.0
psycopg2==2.5.1
=======
Flask-SSLify==0.1.5
gunicorn==19.6.0
psycopg2==2.7.3.2

>>>>>>> 17-app-1
